WL-469: implemented a better strategy for removing middle initial(s).
This implementation is as per Matthew's original suggestion. Here we
extract the first name from the display name by removing the surname. If
the surname is not present - unlikely, but possible - we just resort to the
value of the givenName attribute from Oak LDAP.

WL-652 Don't trust input LDAP data.
Should prevent IOBE with firstname == lastname.