WL-469: implemented alternative strategy for removing middle initial(s).
This is an alternative strategy to the previous one! This method involves
taking the 'displayName' attribute from Oak LDAP and stripping the surname
from it in order to obtain a suitable value for first name. It is thought
that this should be better at coping with names such as 'Jean Luc Goddard'
where Luc is considered part of the first name rather than a middle initial
that needs to be removed!