WL-66 Allow principal to come from config.
Normally we would embed the principal to use when validating a service ticket inside the sakai-jaas.conf file. This worked however as this principal is specific to each service we end up having a sakai-jaas.conf for each deployment. With the switch to docker it would be much cleaner if sakai-jaas.conf was the same everywhere (and embedded in the image) and the principal to use was pulled from the sakai configuration.