WL-4174 Better LDAP connection validator.
We limit the connect timeout when creating new sockets so that when attempting to make a new connection if one of the servers is down we give up after 5 seconds and try another one. Otherwise we only every try one server. We don’t wait forever when attempting to create a new connection as there is a 60 second limit on the pool for creating a new object.

We also limit the lifetime of a LDAP connection to 60 seconds this is for the case when a connections gets shut down with being correctly closed. In this case we will just have to wait for the bad connections to become evicted from the pool.